% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormapping.R
\name{blues}
\alias{blues}
\title{Color-mapping of blue}
\usage{
blues(x)
}
\arguments{
\item{x}{vector to be mapped}
}
\value{
colors
}
\description{
blues
}
\details{
This function shows scale of blue; used to create the legend for rr maps
}
