% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcexpected.R
\name{calcexpected}
\alias{calcexpected}
\title{calcexpected}
\usage{
calcexpected(observed, periods, ids, family = c("poisson", "binomial"))
}
\arguments{
\item{observed}{Vector of observed counts}

\item{periods}{Cector of time periods matching observed counts}

\item{ids}{Vector of centroid identifiers matching the observed counts.}

\item{family}{Must be provided. Can either be specified as \code{poisson} or \code{binomial}.}
}
\value{
returns vector of expected counts
}
\description{
Expected counts can be estimated if not available, following selected distribution function
}
\details{
Estimate expected counts if data not available
}
\examples{
\donttest{
data(japanbreastcancer)
calcexpected(japanbreastcancer$observed, japanbreastcancer$period, japanbreastcancer$id, family="poisson") }
}
