% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusso-package.R, R/clusso.R
\docType{package}
\name{clusso}
\alias{clusso}
\alias{clusso-package}
\title{clusso: Spatial cluster detection using LASSO}
\usage{
clusso(clst, x, y, rMax, Time, utm = TRUE, longdat = TRUE,
  analysis = c("space", "spacetime", "both"), maxclust = 10,
  overdispfloor = TRUE, cv = NULL, collapsetime = FALSE)
}
\arguments{
\item{clst}{list; output from \code{toclusso} function. Must be of class \code{clst}.}

\item{x}{x coordinates (easting/latitude); if utm coordinates, scale to km.}

\item{y}{y coordinates (northing/longitude); if utm coordinates, scale to km.}

\item{rMax}{set max radius (in km).}

\item{Time}{Number of time periods or years in your dataset. Must be declared as numeric.}

\item{utm}{default is \code{TRUE}. If \code{FALSE}, then coordinates will be interpreted as Longitude/Latitude and the haversine formula will be used to determine the distance between points.}

\item{longdat}{Is the data in panel/long format? Default is \code{TRUE}. For wide format, specify \code{FALSE} (TODO).}

\item{analysis}{A string specifying if the spatial (\code{"space"}), spatio-temporal (\code{"spacetime"}), or both spatial and spatio-temporal (\code{"both"}) analysis should be executed. Default is \code{"both"}.}

\item{maxclust}{Upper limit on the maximum number of clusters you expect to find in the region. This equivalent to setting \code{dfmax} in the lasso. If none supplied, default is \code{10}.}

\item{overdispfloor}{overdispfloor default is \code{TRUE}. When TRUE, it limits \eqn{\phi1} (overdispersion parameter) to be greater or equal to 1. If FALSE, will allow for under-dispersion in the model.}

\item{cv}{Numeric argument for the number of folds to use if using k-fold cross-validation. Default is \code{NULL}, indicating that cross-validation should not be performed in favor of \code{clusso}.}

\item{collapsetime}{Default is \code{FALSE}. Alternative definition for space-only model to instead collapse expected and observed counts across time. TODO}
}
\value{
returns list of cluster detection results ready to plot
}
\description{
Runs helper function for both the space and space-time Lasso model on observed data.
}
\details{
Detect a cluster in space or spacetime using Lasso

Set up for clusso
}
\examples{
\donttest{
data(japanbreastcancer)
#Set Some Initial Conditions
x1=utmJapan$utmx/1000
y1=utmJapan$utmy/1000
rMax <- 20 
Time=5
japanbreastcancer <- japanbreastcancer[,-1] #get rid of indicator column
clst <- toclusso(japanbreastcancer, expected = expdeath, observed=death,timeperiod = period)
system.time(resreal <- clusso(clst, x,y, rMax, Time, utm=TRUE, analysis="both", maxclust=10))
 }
}
