% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusso.R
\name{clussoBinom}
\alias{clussoBinom}
\title{clussoBinom}
\usage{
clussoBinom(analysis, x, y, rMax, period, expected, observed, covars, Time,
  utm, maxclust, overdispfloor, cv, collapsetime, nsize)
}
\arguments{
\item{analysis}{A string specifying if the spatial (\code{"space"}), spatio-temporal (\code{"spacetime"}), or both spatial and spatio-temporal (\code{"both"}) analysis should be executed.}

\item{x}{x coordinates (easting/latitude); if utm coordinates, scale to km.}

\item{y}{y coordinates (northing/longitude); if utm coordinates, scale to km.}

\item{rMax}{Set maximum radius for potential clusters (in km).}

\item{period}{Vector of timeperiods in the data set. If this is variable is not a factor in the dataframe, then it will be automatically converted to one by \code{clusso()} with a warning message.}

\item{expected}{Total number of both cases and controls (or number of trials).}

\item{observed}{Vector of number of cases (or successes).}

\item{covars}{Matrix of covariates.}

\item{Time}{Number of timeperiods in the dataset. This is calculated based on the number of unique timeperiods (factor levels) supplied to \code{clusso}.}

\item{utm}{Default is \code{TRUE} (coordinates are in the Universal Transverse Mercator (UTM) coordinate system). If \code{FALSE}, then coordinates will be interpreted as Longitude/Latitude and the Haversine formula will be used to determine the distance between points.}

\item{maxclust}{Upper limit on the maximum number of clusters you expect to find in the region. This equivalent to setting \code{dfmax} in \code{glmnet}. If none supplied, default is \code{11}.}

\item{overdispfloor}{Default is \code{TRUE}. When \code{TRUE}, it limits \eqn{\phi} (overdispersion parameter) to be greater or equal to 1. If \code{FALSE}, it will allow for under-dispersion in the model.}

\item{cv}{Numeric argument for the number of folds to use if using k-fold cross-validation. Default is \code{NULL}, indicating that cross-validation should not be performed in favor of \code{clusso}.}

\item{collapsetime}{Default is \code{FALSE}. Alternative definition for space-only model to instead collapse expected and observed counts across time.}

\item{nsize}{Allows for user-specification of \eqn{n} in information criteria penalty. Default is for finite samples, where in the Poisson case \eqn{n = \mu n} and for the binomial case \eqn{n = min(numcases, numcontrols)}. For the asymptotic case, set to \code{sum(observed)}. Other penalties can also be applied.}
}
\value{
List of lists output from detection.
}
\description{
This function runs both the space and space-time LASSO binomial model. A separate function (clusso) is the helper function which will have
flexibility to specify the space or spacetime or both models to be run .
}
\details{
Detect a cluster in space or spacetime using the LASSO: Binomial model
}
