% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clussoplot.R
\name{clussoplot}
\alias{clussoplot}
\title{clussoplot}
\usage{
clussoplot(outclusso, analysis = c("space", "spacetime", "both"), Time,
  cv = FALSE)
}
\arguments{
\item{outclusso}{outclusso Object with output from \code{clusso}.}

\item{analysis}{A string specifying if the spatial (\code{"space"}), spatio-temporal (\code{"spacetime"}), or both spatial and spatio-temporal (\code{"both"}) analysis should be executed.}

\item{Time}{Number of time periods in the analysis}

\item{cv}{Boolean. If TRUE, then cross-validation was used. Default is FALSE (information criteria used)}
}
\value{
Returns a plot
}
\description{
Plot the coefficient paths from the LASSO-identified potential clusters.
}
\details{
Plot coefficient paths from clusso
}
\examples{
\donttest{
clussoplot(resreal, analysis="both", Time=5)
}
}
