% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clussoplot.R
\name{clussoplotMaster}
\alias{clussoplotMaster}
\title{clussoplotMaster}
\usage{
clussoplotMaster(outclusso, analysistype, Time, maxdim)
}
\arguments{
\item{outclusso}{outclusso Object with output from \code{clusso}.}

\item{Time}{Number of time periods in the analysis}

\item{maxdim}{maximum number of potential clusters}

\item{analysis}{A string specifying if the spatial (\code{"space"}), spatio-temporal (\code{"spacetime"}), or both spatial and spatio-temporal (\code{"both"}) analysis should be executed.}
}
\value{
Returns plots
}
\description{
Plot the coefficient paths from the LASSO-identified potential clusters.
}
\details{
Helper: Plot coefficient paths from clusso
}
