% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clussopretty.R
\name{clussopretty}
\alias{clussopretty}
\title{clussopretty}
\usage{
clussopretty(outclusso, analysis = "both", clusteridentify = FALSE,
  clusterRR)
}
\arguments{
\item{outclusso}{Object with output from \code{clusso}.}

\item{analysis}{A string specifying if the spatial (\code{"space"}), spatio-temporal (\code{"spacetime"}), or both spatial and spatio-temporal (\code{"both"}) analysis should be executed. Default is \code{"both"}.}

\item{clusteridentify}{Whether specific clusters should be identified in output; default is FALSE}

\item{clusterRR}{Risk ratio cut off for a cluster to be identified. Alternatively, a risk ratio can be provided (ex: 1.0, 1.25) that would serve as a cut off value for identifiying a cluster. Current default is 1.0. We define a cluster as being different from the background rate (TODO).}
}
\value{
Data frame of output that can be sent to a csv or list of data frames (if clusteridentify set to TRUE).
}
\description{
This function takes the cluster detection results from the \code{clusso} function and creates a nice table of detected clusters.
}
\details{
Create pretty table from \code{clusso} output
}
\examples{
\donttest{
data(japanbreastcancer)
#Set Some Initial Conditions
x1=utmJapan$utmx/1000
y1=utmJapan$utmy/1000
rMax <- 20 
Time=5
japanbreastcancer <- japanbreastcancer[,-1] #get rid of indicator column
clst <- toclusso(japanbreastcancer, expected = expdeath, observed=death,timeperiod = period)
system.time(outclusso <- clusso(clst, x,y, rMax, Time, utm=TRUE, analysis="both", maxclust=10))
clussopretty(outclusso, analysis="both")
clussopretty(outclusso, analysis="both", clusteridentify=TRUE)
clussopretty(outclusso, analysis="both", clusteridentify=TRUE, clusterRR = 1.05)
 }
}
