% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_sim.R
\name{clust_sim}
\alias{clust_sim}
\title{clust_sim}
\usage{
clust_sim(clst, x, y, rMax, Time, nsim, center, radius, risk.ratio, timeperiod,
  utm = TRUE, byrow = TRUE, threshold, space = c("space", "spacetime",
  "both"), theta = NULL, nullmod = NULL, overdispfloor,
  collapsetime = FALSE, background.rate = NULL)
}
\arguments{
\item{clst}{list; output from toclust function. Must be of class clst.}

\item{x}{x coordinates (easting/latitude); if utm coordinates, scale to km.}

\item{y}{y coordinates (northing/longitude); if utm coordinates, scale to km.}

\item{rMax}{set max radius (in km)}

\item{Time}{Number of time periods or years in your dataset. Must be declared as numeric.}

\item{nsim}{Number of simulations you would like to run}

\item{center}{can be a single center or for multiple clusters, concatenate them. Max three TODO extend this}

\item{radius}{radius for the cluster you want in the simulation}

\item{risk.ratio}{setting for what the risk ratio should be in the cluster to be detected by the simulation}

\item{timeperiod}{time period where the cluster  starts in the simulation
we will only be looking at periods 2 and 5. If multi_period is TRUE, then we will instead consider timeperiod through period_end (timeperiod:period_end). Following the same example,
this would mean we look at periods 2, 3, 4, and 5.}

\item{utm}{default is true}

\item{byrow}{TRUE}

\item{threshold}{vector or value as threshold for cluster detection}

\item{space}{space and space-time. Default is to run all four models: quasi-poisson and poisson for both space and space-time. User can specify, space = space,
space = spacetime, or space = both.}

\item{theta}{default is 1000. Can add in overdispersion to simulated model by changing this value.}

\item{nullmod}{default is NULL; otherwise will run null model}

\item{overdispfloor}{overdispfloor default is TRUE. When TRUE, it limits phi (overdispersion parameter) to be greater or equal to 1. If FALSE, will allow for under dispersion.}

\item{collapsetime}{alternative definition for space-only model to instead collapse expected and observed counts across time. TODO}

\item{background.rate}{option to specify varying background rate instead of varying cluster rate for simulation. Default is null. If this option is specified, risk.ratio must be set to "background".}
}
\value{
returns list of lists
}
\description{
This helper function runs both the space and space-time Lasso model simulations for all 4 models simulataneously: Quasi-Poisson vs. Poisson in both space and space-time.
This function is to be run on simulated data and all four models are run on the same simulated set. 
A separate function (clust.sim) can be used for running simulations on individual models and (clust) can be used for observed data.
}
\examples{
\donttest{
data(japanbreastcancer)
#Set Some Initial Conditions
x1=utmJapan$utmx/1000
y1=utmJapan$utmy/1000
#Set initial parameters
rMax <- 20 
Time=5
nsim=2 #number of simulations
center=1
radius=18
timeperiods = c(1:3)
risk.ratio=2
theta = 10000
overdispfloor = TRUE
nullmod <- TRUE
dframe <- dframe[,-1]
clst <- toclust(japanbreastcancer, expected = japanbreastcancer$expdeath, 
 observed = japanbreastcancer$death,timeperiod = japanbreastcancer$period, covars = FALSE)
res <- clust_sim(clst, x1,y1,rMax, Time, nsim,center, radius, risk.ratio, 
 timeperiod, utm=TRUE, byrow=TRUE, 
threshold, space= "both",theta = theta, nullmod = TRUE, overdispfloor)
}
}

