% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustDiagnostics.R
\name{clustDiagnostics}
\alias{clustDiagnostics}
\title{Suite of function for cluster detection diagnostics - incluster diagnostics}
\usage{
clustDiagnostics(incluster, threshold, nullmod, nsim)
}
\arguments{
\item{incluster}{inherited object from detect_incluster_ic}

\item{threshold}{detection threshold set by user}

\item{nullmod}{default is NULL. If not null, then null model results will be returned}

\item{nsim}{number of simulations}
}
\value{
list of detection
}
\description{
clustDiagnostics
}

