% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters2df.R
\name{clusters2df}
\alias{clusters2df}
\title{Create the clusters dataframe}
\usage{
clusters2df(xP, yP, r.max, utm = FALSE, n)
}
\arguments{
\item{xP}{vector of x coordinates (easting/latitude); if utm coordinates, scale to km.}

\item{yP}{vector of y coordinates (northing/longitude); if utm coordinates, scale to km.}

\item{r.max}{set max radius (in km)}

\item{utm}{TRUE/FALSE as to whether or not the x and y coordinates are in UTM (TRUE) or LAT/LONG(FALSE)}

\item{n}{Number of coordinate pairs/number of centers}
}
\value{
This function returns a dataframe that contains
}
\description{
Create the clusters dataframe
}
\examples{
rMax = 20
#utm example
x_utm <- c(399786.6, 360917.0, 385175.1, 371603.4, 388154.2, 375023.3)
y_utm <- c(4047756, 4023885, 4025749, 4018172, 4047900, 4068053)
#lat/long example
x_latlon <- c(36.569996, 36.350001, 36.370002, 36.299997, 36.570002, 36.749997)
y_latlon<- c(7.88, 7.45, 7.72, 7.57, 7.57, 7.60)
clusters2df(x_utm, y_utm, rMax, utm=TRUE, length(x_utm))
clusters2df(x_latlon, y_latlon, rMax, utm=FALSE, length(x_latlon))
}
