% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.R
\name{clustMaster}
\alias{clustMaster}
\title{clustMaster}
\usage{
clustMaster(analysis, x, y, rMax, period, expected, observed, covars, Time,
  utm, longdat, maxclust, overdispfloor, cv, collapsetime)
}
\arguments{
\item{analysis}{A string specifying if the spatial (\code{"space"}), spatio-temporal (\code{"spacetime"}), or both spatial and spatio-temporal (\code{"both"}) analysis should be executed. Default is \code{"both"}.}

\item{x}{x coordinates (easting/latitude); if utm coordinates, scale to km.}

\item{y}{y coordinates (northing/longitude); if utm coordinates, scale to km.}

\item{rMax}{set max radius (in km)}

\item{period}{vector of periods or years in data set. Should be imported as a factor.}

\item{expected}{vector of expected counts. Expected counts must match up with the year and observed vectors.}

\item{observed}{vector of observed counts. Observed counts must match up with the year and expected vectors.}

\item{covars}{matrix of covariates.}

\item{Time}{Number of time periods or years in your dataset. Must be declared as numeric.}

\item{utm}{default is \code{TRUE}. If \code{FALSE}, then coordinates will be interpreted as Longitude/Latitude and the haversine formula will be used to determine the distance between points.}

\item{longdat}{Is the data in panel/long format? Default is \code{TRUE}. For wide format, specify \code{FALSE} (TODO).}

\item{maxclust}{Upper limit on the maximum number of clusters you expect to find in the region. This equivalent to setting \code{dfmax} in the lasso. If none supplied, default is \code{10}.}

\item{overdispfloor}{overdispfloor default is \code{TRUE}. When TRUE, it limits \eqn{\phi1} (overdispersion parameter) to be greater or equal to 1. If FALSE, will allow for under-dispersion in the model.}

\item{cv}{Numeric argument for the number of folds to use if using k-fold cross-validation. Default is \code{NULL}, indicating that cross-validation should not be performed in favor of \code{clust}.}

\item{collapsetime}{Default is \code{FALSE}. Alternative definition for space-only model to instead collapse expected and observed counts across time. TODO}
}
\value{
list of lists output from detection
}
\description{
This function runs both the space and space-time Lasso model. This function is to be run on observed data. A separate function (clust) is the helper function which will have
flexibility to specify the space or spacetime or both models to be run (TODO).
}
\details{
Detect a cluster in space or spacetime using Lasso on observed data
}
