% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_sim.R
\name{clustMaster_sim}
\alias{clustMaster_sim}
\title{clustMaster_sim}
\usage{
clustMaster_sim(x, y, rMax, period, expected, observed, covars, Time, nsim,
  center, radius, risk.ratio, timeperiod, utm, longdat, thresh,
  theta = theta, nullmod = nullmod, maxclust = maxclust,
  overdispfloor = overdispfloor, collapsetime, background_rate)
}
\arguments{
\item{x}{x coordinates (easting/latitude); if utm coordinates, scale to km.}

\item{y}{y coordinates (northing/longitude); if utm coordinates, scale to km.}

\item{rMax}{set max radius (in km)}

\item{period}{vector of periods or years in dataset. Should be imported as a factor.}

\item{expected}{vector of expected counts. Expected counts must match up with the year and observed vectors.}

\item{observed}{vector of observed counts. Observed counts must match up with the year and expected vectors.}

\item{covars}{dataframe of covariates, if supplied to `toclust` function.}

\item{Time}{Number of time periods or years in your dataset. Must be declared as numeric.}

\item{nsim}{Number of simulations you would like to run}

\item{center}{can be a single center or for multiple clusters, concatenate them. Max three TODO extend this}

\item{radius}{radius for the cluster you want in the simulation}

\item{risk.ratio}{setting for what the risk ratio should be in the cluster to be detected by the simulation}

\item{timeperiod}{time period where the cluster  starts in the simulation
we will only be looking at periods 2 and 5. If multi_period is TRUE, then we will instead consider timeperiod through period_end (timeperiod:period_end). Following the same example,
this would mean we look at periods 2, 3, 4, and 5.}

\item{utm}{utm TRUE/FALSE as to whether or not the x and y coordinates are in UTM (TRUE) or LAT/LONG(FALSE)}

\item{longdat}{longdat default is True. If data should be imported by column then set to FALSE}

\item{thresh}{vector or value as threshold for cluster detection}

\item{theta}{default is 1000. Can add in overdispersion to simulated model by changing this value.}

\item{nullmod}{if TRUE, then null models will be run. Otherwise, default is null.}

\item{maxclust}{Upper limit on the maximum number of clusters you expect to find in the region. This equivalent to setting \code{dfmax} in the lasso. If none supplied, default is \code{10}.}

\item{overdispfloor}{overdispfloor default is TRUE. When TRUE, it limits phi (overdispersion parameter) to be greater or equal to 1. If FALSE, will allow for under dispersion.}

\item{collapsetime}{alternative definition for space-only model to instead collapse expected and observed counts across time. TODO}

\item{background_rate}{option to specify varying background rate instead of varying cluster rate for simulation. Default is null. If this option is specified, risk.ratio must be set to "background".}
}
\value{
returns list of lists
}
\description{
This function runs both the space and space-time Lasso model simulations for all 4 models simulataneously: Quasi-Poisson vs. Poisson in both space and space-time.
This function is to be run on simulated data and all four models are run on the same simulated set. 
A separate function (clust.sim) can be used for running simulations on individual models and (clust) can be used for observed data.
}
