% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbern.R
\name{dbern}
\alias{dbern}
\title{dbern}
\usage{
dbern(Yx, p)
}
\arguments{
\item{Yx}{Cases (0 or 1).}

\item{p}{Vector of estimated probabilities for each individual.}
}
\value{
Returns vector of Bernoulli log-likelihood for each proposed path of the LASSO tuning parameter values.
}
\description{
Bernoullii Function
}
