% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbin.R
\name{dbin}
\alias{dbin}
\title{dbin}
\usage{
dbin(Yx, Ex, p)
}
\arguments{
\item{Yx}{Number of cases.}

\item{Ex}{Total number of trials (or cases + controls).}

\item{p}{Vector of estimated probabilities for each space-time location.}
}
\value{
Returns vector of Binomial log-likelihood for each proposed path of the LASSO tuning parameter values.
}
\description{
Binomial Function
}
