% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_incluster.R
\name{detect_incluster}
\alias{detect_incluster}
\title{detect_incluster}
\usage{
detect_incluster(sparseMAT, lassoresult, vectors.sim, rr, set, timeperiod,
  Time, nsim, under = FALSE, nullmod, risk.ratio, center, radius, x, y,
  rMax, thresh, ncentroids)
}
\arguments{
\item{sparseMAT}{large sparse matrix created in \code{clusso_sim} function.}

\item{lassoresult}{List of QBIC, QAIC, QAICc estimates from the lasso results}

\item{vectors.sim}{dataframe of initial vectors of the observed and expected counts that went into simulation function}

\item{rr}{risk ratio matrix that was used in the simulation}

\item{set}{output from detet_set function}

\item{timeperiod}{Time period span of the cluster}

\item{Time}{number of time periods total in the model}

\item{nsim}{number of simulations run}

\item{under}{Default is FALSE. If TRUE, then it will estimate detection based on a relative risk which is < 1. In other words, it will consider
clusters to be identified where the estimated values are less than the background rate, not more than the background rate as is the case in the
elevated relative risk models.}

\item{nullmod}{Default is NULL. If TRUE, then it will estimate detection based on the null model where there is no cluster.}

\item{risk.ratio}{Risk ratio that was set for cluster in simulation}

\item{center}{center of cluster}

\item{radius}{radius of cluster}

\item{x}{x-coordinates}

\item{y}{y-coordinates}

\item{rMax}{Maximum radius for threshold in simulation}

\item{thresh}{Default is NULL. Vector of thresholds for additional diagnostic criteria for cluster detection.}

\item{ncentroids}{Number of centroids in the data.}
}
\description{
This function will calculate detection based on the three information criterion. You can run detection on a null model (no cluster),
a model with an under-estimated cluster (artificial cluster <1), and on an elevated relative risk cluster. Standard detection criteria include percent of
cells detected that are inside/outside the true cluster and percent of total potential clusters that are at least partially inside/outside the true cluster
that are detected. Additional criteria with threshold options can also be specified.
}
