% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustDiagnostics.R
\name{detect_incluster_ic}
\alias{detect_incluster_ic}
\title{detect_incluster_ic
This function will calculate detection based on the three information criterion. You can run detection on a null model (no cluster),
a model with an under-estimated cluster (artificial cluster <1), and on an elevated relative risk cluster. This is called by the more general
function *detect_incluster* which will switch to this function (TODO - Allow for detection options for AIC, AICc, and BIC only).}
\usage{
detect_incluster_ic(lassoresult, vectors.sim, rr, set, timeperiod, Time, nsim,
  under = FALSE, nullmod)
}
\arguments{
\item{lassoresult}{List of QBIC, QAIC, QAICc estimates from the mylasso.sim function}

\item{vectors.sim}{dataframe of initial vectors of the observed and expected counts that went into simulation function}

\item{rr}{risk ratio matrix that was used in the simulation}

\item{set}{output from detet_set function}

\item{timeperiod}{Time period span of the cluster}

\item{Time}{number of time periods total in the model}

\item{nsim}{number of simulations run}

\item{under}{Default is FALSE. If TRUE, then it will estimate detection based on a relative risk which is < 1. In other words, it will consider 
clusters to be identified where the estimated values are less than the background rate, not more than the background rate as is the case in the 
elevated relative risk models.}

\item{nullmod}{Default is NULL. If TRUE, then it will estimate detection based on the null model where there is no cluster.}
}
\description{
detect_incluster_ic
This function will calculate detection based on the three information criterion. You can run detection on a null model (no cluster),
a model with an under-estimated cluster (artificial cluster <1), and on an elevated relative risk cluster. This is called by the more general
function *detect_incluster* which will switch to this function (TODO - Allow for detection options for AIC, AICc, and BIC only).
}

