% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyplot.R
\name{easyplot}
\alias{easyplot}
\title{Auxiliary Plotting Functions for Japanese Breast Cancer Data}
\usage{
easyplot(prefect, polygons, pdfname, res, mods, space = c("space",
  "spacetime", "both"), obs = NULL)
}
\arguments{
\item{prefect}{prefects dataframe}

\item{polygons}{polygons dataframe}

\item{pdfname}{pdfname of what the output pdf will be called}

\item{res}{resultant list from clust_ function}

\item{mods}{string vector of which models you ran}

\item{space}{If space="space", then the Quasi-Poisson and Poisson spatial only models will be run; if space="spacetime" then the Quasi-Poisson and Poisson
spatio-temporal models will be run; if space="both" then all four models will be run}

\item{obs}{default is NULL. If not null, then will add "observed" instead of "oracle" label to plot for comparison map.}
}
\description{
Wrapper function easyplot
}

