% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_clusteroverlap.R
\name{get_prob}
\alias{get_prob}
\title{#' @title
#' get_prob
#' @description 
#' Finds the probability of being in the cluster for BIC, AIC, and AICc based on the expected risk ratio
#' @param lassoresult result of either space-time lasso simulation or space-only simulation
#' @param init list of initial vector values
#' @param E1 standardized expected counts
#' @param ncentroids number of centroids or centers
#' @param Time number of time periods
#' @param nsim number of simulations
#' @param threshold vector of two threshold values TODO: allow flexibility for number of threshold
#' @return returns list of probabilities.
#'}
\usage{
get_prob(lassoresult, init, E1, ncentroids, Time, nsim, threshold)
}
\description{
#' @title
#' get_prob
#' @description 
#' Finds the probability of being in the cluster for BIC, AIC, and AICc based on the expected risk ratio
#' @param lassoresult result of either space-time lasso simulation or space-only simulation
#' @param init list of initial vector values
#' @param E1 standardized expected counts
#' @param ncentroids number of centroids or centers
#' @param Time number of time periods
#' @param nsim number of simulations
#' @param threshold vector of two threshold values TODO: allow flexibility for number of threshold
#' @return returns list of probabilities.
#'
}
