% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormapping.R
\name{greys}
\alias{greys}
\title{Color-mapping of probabilities to grey scale}
\usage{
greys(x)
}
\arguments{
\item{x}{vectors of probabilities to be mapped}
}
\value{
colors
}
\description{
greys
}
\details{
This function establishes the spread of grey scale for probabilities to be mapped to. Higher probabilities will be darker and lower probabilities will be lighter colors.
}

