% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japan.poly2.R
\docType{data}
\name{japan.poly2}
\alias{japan.poly2}
\title{Polygon borders}
\format{An object of class "data.frame" with 1460 rows and 2 variables:
\itemize{
\item V1: x-coordinate of the vector containing the coordinates of the vertices of the polygon.
\item V2: y-coordinate of the vector containing the coordinates of the vertices of the polygon.
}}
\usage{
japan.poly2
}
\description{
Polygon borders for Japanese Breast Cancer dataset (JBC)
}
\examples{
data(japan.poly2)
}
\keyword{datasets}
