% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdisp.R
\name{overdisp}
\alias{overdisp}
\title{Calculate overdispersion for quasi-likelihood models}
\usage{
overdisp(offset_reg, sim = TRUE, overdispfloor = TRUE)
}
\arguments{
\item{offset_reg}{Object of class glm.}

\item{sim}{For simulation run this takes on \code{TRUE}, else it should be \code{FALSE}. Default is set to \code{TRUE}.}

\item{overdispfloor}{Default is \code{TRUE}. When \code{TRUE}, it limits \eqn{\phi} (overdispersion parameter) to be greater or equal to 1. If \code{FALSE}, it will allow for under-dispersion in the model.}
}
\value{
returns \eqn{\phi} overdispersion parameter.
}
\description{
This function calculates the overdispersion parameter for the QIC 'c' overdispersion parameter.
}
