% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyplot.R
\name{plotmap_bic}
\alias{plotmap_bic}
\title{Space-time plotting}
\usage{
plotmap_bic(prefect, polygons, pdfname, res, rr)
}
\arguments{
\item{prefect}{prefects dataframe}

\item{polygons}{polygons dataframe}

\item{pdfname}{pdfname of what the output pdf will be called}

\item{res}{resultant list from clust_ function}

\item{rr}{if FALSE, will print probability map legend, if TRUE will print legend for risk ratios (redblue scheme)}
}
\description{
Space-time plotting
}
