% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyplot.R
\name{plotmap_S}
\alias{plotmap_S}
\title{Space-only plotting}
\usage{
plotmap_S(prefect, polygons, pdfname, res, obs, rr)
}
\arguments{
\item{prefect}{prefects dataframe}

\item{polygons}{polygons dataframe}

\item{pdfname}{pdfname of what the output pdf will be called}

\item{res}{resultant list from clusso_ function}

\item{obs}{if observed is to be plotted or oracle from simulation}

\item{rr}{is TRUE, result in legend for redblue (risk ratios). Default is grey scale legend}
}
\description{
Space-only plotting
}
