% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyplot.R
\name{plotmap_S_cv}
\alias{plotmap_S_cv}
\title{Space-only plotting}
\usage{
plotmap_S_cv(prefect, polygons, pdfname, res, obs)
}
\arguments{
\item{prefect}{prefects dataframe}

\item{polygons}{polygons dataframe}

\item{pdfname}{pdfname of what the output pdf will be called}

\item{res}{resultant list from clusso_ function}

\item{obs}{if observed is to be plotted or oracle from simulation}
}
\description{
Space-only plotting
}
