% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyplot.R
\name{plotmap_ST_cv}
\alias{plotmap_ST_cv}
\title{Space-time plotting}
\usage{
plotmap_ST_cv(prefect, polygons, pdfname, res, obs, rr)
}
\arguments{
\item{prefect}{prefects dataframe}

\item{polygons}{polygons dataframe}

\item{pdfname}{pdfname of what the output pdf will be called}

\item{res}{resultant list from clust_ function}

\item{obs}{if observed is to be plotted or oracle from simulation}
}
\description{
Space-time plotting
}

