% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_clusteroverlap.R
\name{prob_clusteroverlap}
\alias{prob_clusteroverlap}
\title{prob_clusteroverlap}
\usage{
prob_clusteroverlap(sparseMAT, lassoresult, rr, risk.ratio, x, y, rMax, nsim,
  Time, thresh)
}
\arguments{
\item{sparseMAT}{large sparse matrix created in \code{clust_sim} function.}

\item{lassoresult}{List of QBIC, QAIC, QAICc estimates from the mylasso.sim function}

\item{rr}{risk ratio matrix that was used in the simulation}

\item{risk.ratio}{Risk ratio that was set for cluster in simulation}

\item{x}{x-coordinates}

\item{y}{y-coordinates}

\item{rMax}{Maximum radius for threshold in simulation}

\item{nsim}{number of simulations}

\item{Time}{number of time period}

\item{thresh}{Default is NULL; vector or value as threshold for cluster detection}
}
\value{
returns vector which calculated the number of time the cluster was correctly identified out of the simulations
}
\description{
Finds the probability of any overlap with true cluster based on BIC, AIC, and AICc based on the expected risk ratio
}

