% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_incluster.R
\name{prob_incluster}
\alias{prob_incluster}
\title{prob_incluster}
\usage{
prob_incluster(select_mu, ncentroids, Time, nsim)
}
\arguments{
\item{select_mu}{List of selected mu vectors selected by the respective information criteria.}

\item{ncentroids}{number of centroids}

\item{Time}{number of time period}

\item{nsim}{number of simulations}
}
\value{
returns vector which calculated the number of time the cluster was correctly identified out of the simulations
}
\description{
Mapping colors to create a  probability map. This function will create a probability map based on simulation data. In each simulation, it identifies where a cluster was selected,
compared to the background rate. It then average over the number of simulations, giving us a matrix which ranges from 0 to 1 in probability.
To map this probabilities into a color scheme, please see the $colormapping$ function and select probmap=TRUE. TODO integrate all of this
into a workflow and extend to observed data, not only simulated data.
}

