% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prod_YxCpp}
\alias{prod_YxCpp}
\title{Calculates sequencce of values based on maximum radius}
\usage{
prod_YxCpp(vY, clast, ccenter)
}
\arguments{
\item{vY}{numeric vector}

\item{clast}{last observation in sequence}

\item{ccenter}{centroid value}
}
\value{
numeric vector that feeds into st_matCpp function
}
\description{
helper function
}
