% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormapping.R
\name{redblue}
\alias{redblue}
\title{Color-mapping of risk ratios to red-blue gradient}
\usage{
redblue(x)
}
\arguments{
\item{x}{this will be the risk ratios shrunk to be on the scale of half risk to twice the risk as end points.}
}
\value{
colors
}
\description{
This function establishes the spread of reds and blues for the risk ratios to be mapped to. Higher risk ratios will be deeper red colors and lower risk ratios will be deeper blue colors.
}

