% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormapping.R
\name{reds}
\alias{reds}
\title{Color-mapping of probabilities to red scale}
\usage{
reds(x)
}
\arguments{
\item{x}{this will be the risk ratios shrunk to be on the scale of half risk to twice the risk as end points.}
}
\value{
colors
}
\description{
reds
}
\details{
This function establishes the spread of grey scale for probabilities to be mapped to. Higher probabilities will be darker and lower probabilities will be lighter colors.
}

