% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale}
\alias{scale}
\title{Re-scaling data to get SMR (standardized mortality rate)}
\usage{
scale(init, Time)
}
\arguments{
\item{init}{list of initial settings in spacetimeLasso}

\item{Time}{number of time periods}
}
\description{
This function standardizes the fitted values by the sum of the observed/the sum of the expected within each time period. Inputs should all be in vector form.
The function handles standardizing within time period as long.
}
\examples{

#set up
set.seed(2)
Time <- 2
theta = 1000
nsim = 2
period <- rep(seq(1,2),5)
expected <- MASS::rnegbin(n = 10,mu = 15,theta = 1000)
observed <- MASS::rnegbin(expected, theta=1000)
covars <- NULL
#set init
init <- setVectors(period, expected, observed, covars, Time, byrow=TRUE)
scale(init, Time)
}
