% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_sim}
\alias{scale_sim}
\title{Re-scaling data to get SMR (standardized mortality rate) for simulation}
\usage{
scale_sim(YSIM, init, nsim, Time)
}
\arguments{
\item{YSIM}{list of simulated observed}

\item{init}{list of initial settings in spacetimeLasso}

\item{nsim}{number of simulations}

\item{Time}{number of time periods}
}
\description{
This function standardizes the fitted values by the sum of the *simulated* observed/the sum of the expected within each time period. Inputs should all be in vector form.
The function handles standardizing within time period as long. This function should be used with simulation study.
}

