% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVectors.R
\name{setVectors}
\alias{setVectors}
\title{setVectors}
\usage{
setVectors(period, expect, observed, covars, Time, byrow = TRUE)
}
\arguments{
\item{period}{Vector of timeperiods in the data set. If this is variable is not a factor in the dataframe, then it will be automatically converted to one by \code{clusso()} with a warning message. Periods must match up with observed and expected vectors.}

\item{expect}{Vector of expected counts. Expected counts must match up with period and observed vectors.}

\item{observed}{Vector of observed counts. Observed counts must match up with period and expected vectors.}

\item{covars}{Dataframe of covariates. \code{NULL} if no covariates supplied.}

\item{Time}{Number of timeperiods in the dataset. This is calculated based on the number of unique timeperiods (factor levels) supplied to \code{clusso}.}

\item{byrow}{Set to \code{TRUE}. Data from the dataset should be imported by row, meaning that data must be sorted by geographic identifier, with repeated measurements for each geographic unit repeated.}
}
\value{
This function returns a list of expected and observed counts along with the period.
}
\description{
Creates a List Arranged by Time Period with Expected and Observed Counts and Time Period
}
\examples{
period1 <- rep(c("1","2"),times=5)
expected <- MASS::rnegbin(n = 10,mu = 15,theta = 1000)
observed <- MASS::rnegbin(expected, theta=1000)
Time = 2
covars <- NULL
setVectors(period1, expected, observed, covars, Time)
}
