% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimeLasso.R
\name{spacetimeLasso}
\alias{spacetimeLasso}
\title{spacetimeLasso}
\usage{
spacetimeLasso(sparseMAT, n_uniq, vectors, Time, spacetime = TRUE,
  pois = FALSE, maxclust, overdispfloor, cv)
}
\arguments{
\item{sparseMAT}{large sparse matrix created in \code{clust} function.}

\item{n_uniq}{number of unique polygons (ex: counties, zip code, etc). Inherited from \code{clust}.}

\item{vectors}{takes in the list of expected and observed counts from setVectors function}

\item{Time}{number of time periods in the dataset}

\item{spacetime}{indicator of whether the cluster detection method should be run on all space-time clusters(default) or on only the potential space clusters.}

\item{pois}{whether or not the Quasi-Poisson or Poisson model should be run. Default is pois=FALSE (default is Quasi-Poisson model is to be run)}

\item{maxclust}{Upper limit on the maximum number of clusters you expect to find in the region. This equivalent to setting \code{dfmax} in the lasso. If none supplied, default is \code{10}.}

\item{overdispfloor}{default is TRUE. If TRUE, does not allow for underdispersion. If FALSE, allows for underdispersion (phi < 1)}

\item{cv}{option for cross-validation instead of AIC/BIC. Default is set to FALSE}
}
\value{
This function will return a list with the expected counts as selected by QBIC, QAIC, QAICc, a list of original expected counts (Ex),
a list of observed counts (Yx), the lasso object, a list of K values (number of unique values in each decision path), and n (length of unique centers in the clusters dataframe)
}
\description{
This function runs the Lasso regularization technique on our large sparse matric of potential space or space-time clusters.
}
\details{
Spatial and Spatio-Temporal Lasso
}
