% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimeLasso.R
\name{spacetimeLassoPois}
\alias{spacetimeLassoPois}
\title{spacetimeLassoPois}
\usage{
spacetimeLassoPois(lasso, coefs.lasso.all, loglike, mu, K, quasi, covars,
  Yx, Ex, Period, Time, n_uniq, overdispfloor, maxclust, nsize)
}
\arguments{
\item{lasso}{Output from \code{glmnet}.}

\item{coefs.lasso.all}{Matrix of coefficient estimates for every lambda in lasso path.}

\item{loglike}{Loglikelihood for Poisson model.}

\item{mu}{Estimated relative risk estimates (\eqn{\rho_i E_i}).}

\item{K}{Vector of the number of K parameters estimated for every lambda in lasso path.}

\item{quasi}{Boolean. \code{TRUE} indicates a quasi-Poisson model that accounts for overdispersion. \code{FALSE} indicates a Poisson model without adjustment for overdispersion.}

\item{covars}{Dataframe of additional covariates to be included in the model that are un-penalized by the LASSO.}

\item{Yx}{Number of observed cases for each space-time location.}

\item{Ex}{Expected number of cases for each space-time location.}

\item{Period}{Vector of time periods for each space-time location.}

\item{Time}{Integer. Number of time periods in the study.}

\item{n_uniq}{Number of unique centroids in the study area.}

\item{overdispfloor}{Default is \code{TRUE}. When TRUE, it limits \eqn{\phi} (overdispersion parameter) to be greater or equal to 1. If FALSE, will allow for under-dispersion in the model.}

\item{maxclust}{Upper limit on the maximum number of clusters you expect to find in the region. This equivalent to setting \code{dfmax} in the lasso. If none supplied, default is \code{11}.}

\item{nsize}{Allows for user-specification of \eqn{n} in information criteria penalty. Default is for finite samples, where in the Poisson case \eqn{n = \mu n} and for the binomial case \eqn{n = min(numcases, numcontrols)}. For the asymptotic case, set to \code{sum(observed)}. Other penalties can also be applied.}
}
\value{
List of results from Poisson/quasi-Poisson models.
}
\description{
spacetimeLassoPois
}
