% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimeMat.R
\name{spacetimeMat}
\alias{spacetimeMat}
\title{Spatio-temporal Sparse Matrix}
\usage{
spacetimeMat(clusters, numCenters, Time)
}
\arguments{
\item{clusters}{Clusters dataframe that includes the center, x,y, r (radius), n (counter), and last (last observation in potential cluster).}

\item{numCenters}{Number of geographic regions/centroids.}

\item{Time}{Number of timeperiods in the dataset. This is calculated based on the number of unique timeperiods (factor levels) supplied to \code{clusso}.}
}
\value{
Returns sparse space time matrix.
}
\description{
This function takes the Kronecker product of the space and time matrices to create the space-time matrix.
}
