% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{st_matCpp}
\alias{st_matCpp}
\title{Helper function}
\usage{
st_matCpp(X, last, center, T)
}
\arguments{
\item{X}{takes in numeric matrix}

\item{last}{takes in integer vector of last observation in a sequence given a radius}

\item{center}{centroid value}

\item{T}{integer}
}
\value{
numeric vector
}
\description{
helper function
}

