% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimeLasso.R
\name{stLasso_cv}
\alias{stLasso_cv}
\title{stLasso_cv}
\usage{
stLasso_cv(lasso, sparseMAT, Ex, Yx)
}
\arguments{
\item{lasso}{results of cv.glmnet}

\item{sparseMAT}{sparse spatial or spatio-temporal design matrix}

\item{Ex}{expected counts (inherits from setVectors)}

\item{Yx}{observed counts (inherits from setVectors)}
}
\value{
list of expected values, number of clusters, Ex, Yx, and lasso object
}
\description{
This function will output results from cross-validation results.
}
\details{
Selection via cross-validation
}
