% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimeMat.R
\name{timeMat}
\alias{timeMat}
\title{timeMat}
\usage{
timeMat(Time)
}
\arguments{
\item{Time}{Number of time periods in the data}
}
\value{
returns sparse matrix of 1's as indicators of membership in the time cluster
}
\description{
This function creates a sparse matrix of 1's of all of the potential time periods for the cluster to be in. The number of
potential time periods is determined by \link{(Time*(Time-1))}/2.
}
