% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toclust.R
\name{toclust}
\alias{toclust}
\title{toclust

Creates \code{clst} object.}
\usage{
toclust(df, expected, observed, timeperiod, covars = FALSE)
}
\arguments{
\item{df}{name of dataframe.}

\item{expected}{Name of variable that contains the expected counts.}

\item{observed}{Name of variable that contains the observed counts.}

\item{timeperiod}{Name of variable that contains the timeperiod in which counts were observed (as factor). 
If spatial-only analysis, create a column that has a single value (ex: "Time1") and convert this to a factor.}

\item{covars}{are there additional covariates in the dataframe beyond the three required? If so, set to \code{TRUE}. Default is \code{FALSE}.}
}
\value{
\code{clst} object
}
\description{
Creates \code{clst} object for creating potential spatial and spatio-temporal clusters. This function creates a \code{clst} 
object which will contain the expected, observed, and timeperiod information necessary to run \pkg{clust}.
If dataframe is fed in, assumes panel format - see \code{vignette} for details.
}
\examples{
\donttest{
data(japanbreastcancer)
clst <- toclust(japanbreastcancer, expected = expdeath, observed=death,timeperiod = period, covars = FALSE)  
}
}
