% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors_space.R
\name{vectors_space}
\alias{vectors_space}
\title{vectors_space}
\usage{
vectors_space(x, Ex, Yx, Time, init)
}
\arguments{
\item{x}{vector coordinates (unique regardless of time period)}

\item{Ex}{list of simulated and standardized expected counts}

\item{Yx}{observed}

\item{Time}{number of time periods}

\item{init}{initial list of vectors, inherited from function \code{setVectors}.}
}
\value{
returns spatial units with counts aggregated over time periods.
}
\description{
This function will collapse a space-time vector onto space only
}
\details{
Collapse space-tie onto space by summing counts over time by spatial unit.
}
