% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_sim.R
\name{vectors_space_sim}
\alias{vectors_space_sim}
\title{vectors_space_sim}
\usage{
vectors_space_sim(x, Ex, YSIM, Time, init)
}
\arguments{
\item{x}{vector coordinates (unique regardless of time period)}

\item{Ex}{list of simulated and standardized expected counts}

\item{YSIM}{simulated observed}

\item{Time}{number of time periods}

\item{init}{initial list of vectors, inherited from function setVectors.}
}
\value{
returns space-time
}
\description{
Simulation functions for space and space-time cluster detection with lasso. This function will collapse a space-time vector onto space only
}

