% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcexpected.R
\name{calcexpected}
\alias{calcexpected}
\title{calcexpected}
\usage{
calcexpected(observed, population, periods, ids)
}
\arguments{
\item{observed}{Vector of observed counts.}

\item{population}{Vector of population sizes in each geographic unit.}

\item{periods}{Vector of time periods matching observed counts.}

\item{ids}{Vector of centroid identifiers matching the observed counts.}
}
\value{
Returns a dataframe of geographic ids, time periods, population counts, observed counts, and expected counts.
}
\description{
Expected counts can be estimated if not available, following selected distribution function
}
\details{
Estimate expected counts if data not available
}
\examples{
\donttest{
data(japanbreastcancer)
calcexpected(japanbreastcancer$observed, japanbreastcancer$period, japanbreastcancer$id, family="poisson") }
}
