% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcexpected.R
\name{calcexpected}
\alias{calcexpected}
\title{calcexpected}
\usage{
calcexpected(observed, population, periods, ids)
}
\arguments{
\item{observed}{Vector of observed counts.}

\item{population}{Vector of population sizes in each geographic unit.}

\item{periods}{Vector of time periods matching observed counts.}

\item{ids}{Vector of centroid identifiers matching the observed counts.}
}
\value{
Returns a dataframe of geographic ids, time periods, population counts, observed counts, and expected counts.
}
\description{
Expected counts can be estimated if not available, following selected distribution function
}
\details{
Estimate expected counts if data not available
}
\examples{
\donttest{
set.seed(2)
period <- rep(c("1","2"), times=5)
observed <- MASS::rnegbin(n = length(period), mu=20, theta=2)
pop <-  MASS::rnegbin(n = length(period), mu=200, theta=2)
ids <- rep(1:5, each=2)
calcexpected(observed, pop, period, ids) }
}
