% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clussoplot.R
\name{clussoplot}
\alias{clussoplot}
\title{clussoplot}
\usage{
clussoplot(
  outclusso,
  analysis = c("space", "spacetime", "both"),
  model = c("poisson", "binomial"),
  Time,
  collapsetime = FALSE,
  cv = FALSE
)
}
\arguments{
\item{outclusso}{outclusso Object with output from \code{clusso}.}

\item{analysis}{A string specifying if the spatial (\code{"space"}), spatio-temporal (\code{"spacetime"}), or both spatial and spatio-temporal (\code{"both"}) analysis should be executed.}

\item{model}{A string specifying which model to use, Poisson or binomial. For Poisson, specify \code{"poisson"} and both the Poisson and quasi-Poisson model results are returned. For binomial, specify \code{"binomial"} and both the binomial and quasi-binomial model results are returned.}

\item{Time}{Number of time periods in the analysis.}

\item{collapsetime}{Default is \code{FALSE}. Alternative definition for space-only model to instead collapse expected and observed counts across time by summing the counts.}

\item{cv}{Boolean. Takes \code{TRUE} if cross-validation was used. Default is \code{FALSE}.}
}
\value{
Returns a ggplot.
}
\description{
Plot the coefficient paths from the LASSO-identified potential clusters.
}
\details{
Plot coefficient paths from clusso
}
