% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clussoplot.R
\name{clussoplotIC}
\alias{clussoplotIC}
\title{clussoplotIC}
\usage{
clussoplotIC(outclusso, analysistype, model, Time, maxdim)
}
\arguments{
\item{outclusso}{outclusso Object with output from \code{clusso}.}

\item{analysistype}{A string specifying if the spatial (\code{"space"}), spatio-temporal (\code{"spacetime"}), or both spatial and spatio-temporal (\code{"both"}) analysis should be executed.}

\item{model}{A string specifying which model to use, Poisson or binomial. For Poisson, specify \code{"poisson"} and both the Poisson and quasi-Poisson model results are returned. For binomial, specify \code{"binomial"} and both the binomial and quasi-binomial model results are returned.}

\item{Time}{Number of time periods in the analysis.}

\item{maxdim}{Maximum number of potential clusters.}
}
\value{
Returns plots based on information criteria.
}
\description{
Plot the coefficient paths from the LASSO-identified potential clusters (based on information criteria).
}
\details{
Helper: Plot coefficient paths (based on information criteria) from clusso
}
