% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.R
\name{clust}
\alias{clust}
\title{clust}
\usage{
clust(clst, x, y, rMax, Time, utm = TRUE, byrow = TRUE, space = c("space",
  "spacetime", "both"), overdispfloor = NULL, cv = NULL)
}
\arguments{
\item{clst}{list; output from toclust function. Must be of class clst.}

\item{x}{x coordinates (easting/latitude); if utm coordinates, scale to km.}

\item{y}{y coordinates (northing/longitude); if utm coordinates, scale to km.}

\item{rMax}{set max radius (in km)}

\item{Time}{Number of time periods or years in your dataset. Must be declared as numeric.}

\item{utm}{default is TRUE. If FALSE, then will run long/lat data}

\item{byrow}{default is True. If data should be imported by column then set to FALSE}

\item{space}{space and space-time. Default is to run all four models: quasi-poisson and poisson for both space and space-time. User can specify, space = space,
space = spacetime, or space = both.}

\item{overdispfloor}{overdispfloor default is TRUE. When TRUE, it limits phi (overdispersion parameter) to be greater or equal to 1. If FALSE, will allow for under dispersion.}

\item{cv}{option for cross-validation}
}
\value{
returns list of cluster detection results ready to plot
}
\description{
Runs helper function for both the space and space-time Lasso model on observed data. A separate function (clust_sim) can be used for simulating data and running diagnostics on simulations.
}
\details{
Set up for clust
}
\examples{
data(japanbreastcancer)
#Set Some Initial Conditions
x1=utmJapan$utmx/1000
y1=utmJapan$utmy/1000
rMax <- 20 
Time=5
japanbreastcancer <- japanbreastcancer[,-1] #get rid of indicator column
clst <- toclust(japanbreastcancer, expected = japanbreastcancer$expdeath, 
 observed = japanbreastcancer$death,timeperiod = japanbreastcancer$period, covars = FALSE)
system.time(res <- clust(clst, x1,y1, rMax, Time, utm=TRUE, byrow=TRUE, 
 space="both", overdispfloor=TRUE, cv=NULL))
}

