% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormapping.R
\name{colormapping}
\alias{colormapping}
\title{colormapping}
\usage{
colormapping(rate, Time, cv, prob)
}
\arguments{
\item{rate}{this will be the risk ratios shrunk to be on the scale of half risk to twice the risk as end points.}

\item{Time}{how many time periods are in the model? If this is only a spatial model, then time is set to 1}

\item{cv}{triggered if cross-validation model run on real data}

\item{prob}{TRUE or FALSE depending on if you want to map probabilities or not (default is to map RR)}
}
\value{
returns vectors ofcolors for each time period, where risk ratios have been constrained to be between half risk and twice the risk
}
\description{
This function establishes the spread of reds and blues for the risk ratios to be mapped to. Higher risk ratios will be deeper red colors and lower risk ratios will be deeper blue colors.
}

