% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyplot.R
\name{easyplot}
\alias{easyplot}
\title{Auxiliary Plotting Functions for Japanese Breast Cancer Data}
\usage{
easyplot(prefect, polygons, pdfname, rescols, mods, space = c("space",
  "spacetime", "both", "BIC"), probmap, cv = NULL, obs = NULL, rr)
}
\arguments{
\item{prefect}{prefects dataframe}

\item{polygons}{polygons dataframe}

\item{pdfname}{pdfname of what the output pdf will be called}

\item{rescols}{result colors to plot}

\item{mods}{string vector of which models you ran}

\item{space}{If space="space", then the Quasi-Poisson and Poisson spatial only models will be run; if space="spacetime" then the Quasi-Poisson and Poisson
spatio-temporal models will be run; if space="both" then all four models will be run}

\item{probmap}{TRUE or FALSE if probabilities should be mapped or relative risks}

\item{cv}{default is NULL. If not null, then results from cross-validation will be plotted (for real data example)}

\item{obs}{default is NULL. If not null, then will add "observed" instead of "oracle" label to plot for comparison map.}

\item{rr}{TRUE or FALSE. TRUE indicates risk ratios should be plotted, FALSE indicates probabilities should be plotted.}
}
\description{
Wrapper function easyplot
}
\examples{
\donttest{
mods <- c("QuasiPoisson", "Poisson")
pdfname1 <- paste0("rrmaps.pdf")
pdfname2 <- paste0("probabilitymaps.pdf")
easyplot(prefect = japan.prefect2 , polygons = japan.poly2 ,pdfname1,
 res$rrcolors, mods, space="both", probmap=FALSE, obs = NULL, rr=TRUE)
easyplot(prefect = japan.prefect2 , polygons = japan.poly2 ,pdfname2, 
 res$probcolors, mods, space="both", probmap=TRUE, obs = NULL,rr=FALSE)
}
}

