% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_clusteroverlap.R
\name{get_prob}
\alias{get_prob}
\title{get_prob}
\usage{
get_prob(lassoresult, init, E1, ncentroids, Time, nsim, threshold)
}
\arguments{
\item{lassoresult}{result of either space-time lasso simulation or space-only simulation}

\item{init}{list of initial vector values}

\item{E1}{standardized expected counts}

\item{ncentroids}{number of centroids or centers}

\item{Time}{number of time periods}

\item{nsim}{number of simulations}

\item{threshold}{vector of two threshold values TODO: allow flexibility for number of threshold}
}
\value{
returns list of probabilities.
}
\description{
Finds the probability of being in the cluster for BIC, AIC, and AICc based on the expected risk ratio
}
