% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japanbreastcancer.R
\docType{data}
\name{japanbreastcancer}
\alias{japanbreastcancer}
\title{Japanese Breast Cancer Data}
\format{An object of class \code("data.frame"). Includes four separate dataframes: 
\itemize{
 \item 1) dframe.poly2 - coordinates for plotting polygons
 \item 2) dframe.prefect2 - coordinates for plotting prefect boundaries
 \item 3) japanbreastcancer - data on observed death from breast cancer and expected death by polygon-time period.
 \item 4) utmJapan - UTM coordinates
}}
\usage{
data(japanbreastcancer)
}
\description{
Data on breast cancer incidence in prefects in Japan. UTM coordinates and polygon data included.
}
\examples{
data(japanbreastcancer)
}
\references{

}
\keyword{datasets}

