% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdisp.R
\name{overdisp}
\alias{overdisp}
\title{Calculate overdispersion for quasi-likelihood models}
\usage{
overdisp(offset_reg, sim = TRUE, overdispfloor = TRUE)
}
\arguments{
\item{offset_reg}{object of class glm}

\item{sim}{for a simulation run this takes on TRUE, else it should be FALSE. Default is set to TRUE.}

\item{overdispfloor}{Default is set to TRUE so model will not estimate any under-dispersion where phi < 1. The overdispfloor is set for phi to be 1. In a case where this}
}
\value{
returns phi overdispersion parameter
}
\description{
This function calculates the overdispersion parameter for the QIC 'c' overdispersion parameter.
}

