% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbin.R
\name{pihat}
\alias{pihat}
\title{pihat}
\usage{
pihat(xbeta)
}
\arguments{
\item{xbeta}{Sparse matrix of potential space-time clusters times coefficient for each lasso path.}
}
\value{
Returns matrix of estimated probabilities for each space-time location by each lasso path.
}
\description{
Convert linear predictor to probability scale
}
