% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_clusteroverlap.R
\name{prob_clusteroverlap}
\alias{prob_clusteroverlap}
\title{prob_clusteroverlap}
\usage{
prob_clusteroverlap(sparseMAT, lassoresult, rr, risk.ratio, x, y, rMax,
  nsim, Time, thresh, ncentroids, nullmod = NULL)
}
\arguments{
\item{sparseMAT}{large sparse matrix created in \code{clusso_sim} function.}

\item{lassoresult}{List of QBIC, QAIC, QAICc estimates from the mylasso.sim function}

\item{rr}{risk ratio matrix that was used in the simulation}

\item{risk.ratio}{Risk ratio that was set for cluster in simulation}

\item{x}{x-coordinates}

\item{y}{y-coordinates}

\item{rMax}{Maximum radius for threshold in simulation}

\item{nsim}{number of simulations}

\item{Time}{number of time period}

\item{thresh}{thresholds for simulation detection (TODO)}

\item{ncentroids}{number of centroids}

\item{nullmod}{Is model run the the null model with no cluster to detect? Default is that model has a cluster to detect (\code{nullmod=NULL})}
}
\value{
List of detection probabilities (in vs. out of cluster) by criterion. Returns vector which calculated the number of time the cluster was correctly identified out of the simulations.
}
\description{
Finds the probability of any overlap with true cluster based on BIC, AIC, and AICc based on the expected risk ratio
}
