% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormapping.R
\name{reds}
\alias{reds}
\title{Color-mapping of red}
\usage{
reds(x)
}
\arguments{
\item{x}{Vector of probabilities to be mapped.}
}
\value{
colors
}
\description{
reds
}
\details{
This function shows scale of red; used to create the legend for rr maps
}
