% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceMat.R
\name{spaceMat}
\alias{spaceMat}
\title{Creating sparse spatial matrix based on clusters dataframe}
\usage{
spaceMat(clusters, numCenters)
}
\arguments{
\item{clusters}{clusters dataframe that includes the center, x,y, r (radius), n (counter), and last (last observation in potential cluster)}

\item{numCenters}{the number of centers/centroids}
}
\value{
returns sparse matrix of 1's
}
\description{
This function creates a sparse matrix of 1's of all potential clusters for the Lasso algorithm to cycle over; this will take in the potential clusters dataframe
created by \code{clusters2df()} function
}
