% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimeLasso.R
\name{spacetimeLasso}
\alias{spacetimeLasso}
\title{spacetimeLasso}
\usage{
spacetimeLasso(clusters, vectors, Time, spacetime = TRUE, pois = FALSE,
  overdispfloor, cv)
}
\arguments{
\item{clusters}{clusters dataframe from (cluster.df function) that includes the center, x,y, r (radius), n (counter), and last (last observation in potential cluster)}

\item{vectors}{takes in the list of expected and observed counts from setVectors function}

\item{Time}{number of time periods in the dataset}

\item{spacetime}{indicator of whether the cluster detection method should be run on all space-time clusters(default) or on only the potential space clusters.}

\item{pois}{whether or not the Quasi-Poisson or Poisson model should be run. Default is pois=FALSE (default is Quasi-Poisson model is to be run)}

\item{overdispfloor}{default is TRUE. If TRUE, does not allow for underdispersion. If FALSE, allows for underdispersion (phi < 1)}

\item{cv}{option for cross-validation instead of AIC/BIC. Default is set to FALSE}
}
\value{
This function will return a list with the expected counts as selected by QBIC, QAIC, QAICc, a list of original expected counts (Ex),
a list of observed counts (Yx), the lasso object, a list of K values (number of unique values in each decision path), and n (length of unique centers in the clusters dataframe)
}
\description{
This function runs the Lasso regularization technique on our large sparse matric of potential space or space-time clusters.
}
\details{
Spatial and Spatio-Temporal Lasso
}

