% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimeLasso_sim.R
\name{spacetimeLasso_sim}
\alias{spacetimeLasso_sim}
\title{Simulations for Spatial and Spatio-Temporal Lasso}
\usage{
spacetimeLasso_sim(sparseMAT, n_uniq, vectors.sim, Time, spacetime, pois, nsim,
  YSIM, overdispfloor)
}
\arguments{
\item{sparseMAT}{large sparse matrix created in \code{clust_sim} function.}

\item{n_uniq}{number of unique polygons (ex: counties, zip code, etc). Inherited from \code{clust_sim}.}

\item{vectors.sim}{takes in the list of expected and observed counts from setVectors function}

\item{Time}{number of time periods in the dataset}

\item{spacetime}{indicator of whether the cluster detection method should be run on all space-time clusters(default) or on only the potential space clusters.}

\item{pois}{whether or not the Quasi-Poisson or Poisson model should be run. Default is pois=FALSE (default is Quasi-Poisson model is to be run)}

\item{nsim}{number of simulations}

\item{YSIM}{vector of simulated observed counts}

\item{overdispfloor}{default is TRUE. When TRUE, it limits phi (overdispersion parameter) to be greater or equal to 1. If FALSE, will allow for under dispersion.}
}
\value{
This function will return a list with the expected counts as selected by QBIC, QAIC, QAICc, a list of original expected counts (Ex),
a list of observed counts (Yx), the lasso object, a list of K values (number of unique values in each decision path), and n (length of unique centers in the clusters dataframe)
}
\description{
This function runs the Lasso regularization technique on our large sparse matric of potential space-time clusters.It is specifically created to use with simulations.
}

