% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimeMat.R
\name{spacetimeMat}
\alias{spacetimeMat}
\title{Spatio-temporal Sparse Matrix}
\usage{
spacetimeMat(clusters, numCenters, Time)
}
\arguments{
\item{clusters}{the potential clusters dataframe created by \emph{clusters2df}}

\item{numCenters}{number of centroids in space matrix}

\item{Time}{number of time periods}
}
\value{
Returns sparse space time matrix
}
\description{
This function takes the Kronecker product of the space and time matrices to create the space-time matrix
}
