% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetimeMat.R
\name{timeMat}
\alias{timeMat}
\title{timeMat}
\usage{
timeMat(Time)
}
\arguments{
\item{Time}{Number of timeperiods in the dataset. This is calculated based on the number of unique timeperiods (factor levels) supplied to \code{clusso}.}
}
\value{
Returns sparse matrix of 1's as indicators of membership in the time dimension.
}
\description{
This function creates a sparse matrix of 1's indicating  all of the time periods a potential cluster could be in. The number of
potential time periods is determined by \code{[(Time*(Time-1))]/2}.
}
