% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toclust.R
\name{toclust}
\alias{toclust}
\title{toclust}
\usage{
toclust(df, expected, observed, timeperiod, covars = FALSE)
}
\arguments{
\item{df}{dataframe of variables}

\item{expected}{vector or column in a dataframe of expected counts. Format must be supplied as df$var}

\item{observed}{vector or column in a dataframe of observed counts. Format must be supplied as df$var}

\item{timeperiod}{vector or column in a dataframe of timeperiod (should be converted to a factor beforehand). Format must be supplied as df$var}

\item{covars}{are there additional covariates in the dataframe beyond the three required? Default is FALSE}
}
\value{
clst object
}
\description{
Create clst object for creating potential spatial and spatio-temporal clusters. This function will create a clst object which will contain the expected, observed, and time period information necessary to run further cluster functions.
If dataframe is fed in, assumes panel format. Also used for converting covariates to proper object.
}

